Current chain box/button exported into their separate components
